backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
aiofiles==23.2.0
httpx==0.25.2
python-dotenv==1.0.0
apscheduler==3.10.4
pydantic==2.5.0
python-multipart==0.0.6
